move debug coord transform to scanning logic and use a real rotation
